+++ Core/SavedVariables.lua
@@ -1,7 +1,7 @@
 local addonName, SM = ...
 
 local defaults = {
-    version = "1.4.0",
+    version = "1.5.0",
     selectedQuestline = 1,
     viewedLoreChapters = {},
     playedChapters = {},